Create validateUserSession - Validates wos-session and finds corresponding user in database
